Unparsing (serialising) channel and net client
The unparsing is not quite symmetric to the parsing: it doesn't try to
batch writes, so each term sent is sent as a buffer (maybe netty
batches?). The concat'ed-string-to-buffer is good enough for now.

Setting up the channels in net/create-client is slightly mind-bending because of the promise-y nature of the net channel; and, I'm not sure I need the channel pair.