Intermediate between the client and netty with async/chan
With the simple machine: when you read something, put it on <in>, when
there's something on <out>, write it. All ops done in go blocks, with
the idea that they can be parked if there's some other work to do.

The next step is to take notice of backpressure: if you can't write,
stop reading.