Parse incoming buffers and write result back
.. just to test the parser.

A thing I don't like is that the client loop has to know it's getting
`ByteBuf`s to do its inner loop over the bytes; I'd rather do that in
the reader, but it's harder to maintain state without that outer
loop. I guess it could be done in an atom, even though it is needless.